chore(storage, provider): rename bundle state with receipts to execution outcome
