+++ crates/storage/provider/src/bundle_state/mod.rs
@@ -1,13 +1,12 @@
 //! Bundle state module.
 //! This module contains all the logic related to bundle state.
-mod bundle_state_with_receipts;
+
+mod execution_outcome;
 mod hashed_state_changes;
 mod state_changes;
 mod state_reverts;
 
-pub use bundle_state_with_receipts::{
-    AccountRevertInit, BundleStateInit, OriginalValuesKnown, RevertsInit,
-};
+pub use execution_outcome::{AccountRevertInit, BundleStateInit, OriginalValuesKnown, RevertsInit};
 pub use hashed_state_changes::HashedStateChanges;
 pub use state_changes::StateChanges;
 pub use state_reverts::{StateReverts, StorageRevertsIter};
